Update to serde 0.9, add explicit codecov.yml
Now that rust 1.15 is here we can make use of serde's custom derive
in stable. There is no longer any need for us to have the wacky build.rs
to do codegen.

We also add a codecov.yml to control the ranges of coverage considered
acceptable for this project.
