+++ codecov.yml
@@ -1,0 +1,22 @@
+coverage:
+  precision: 2
+  round: down
+  range: 70...100
+
+  status:
+    # Learn more at http://docs.codecov.io/docs/codecov-yaml
+    changes: false
+
+    patch:
+      default:
+        target: 99%
+        threshold: 1%
+
+    project:
+      default:
+        target: 94%
+        threshold: 1%
+        
+comment:
+  layout: "header, diff"
+  behavior: default  # update if exists else create new
